+++ Journal/Journal/Assets.xcassets/AppIcon.appiconset/Contents.json
