+++ Journal/Journal/Assets.xcassets/Contents.json
