+++ Journal/Journal/Base.lproj/LaunchScreen.storyboard
