+++ Journal/Journal/EntryDetailViewController.swift
@@ -1,0 +1,42 @@
+//
+//  EntryDetailViewController.swift
+//  Journal
+//
+//  Created by Tobi Kuyoro on 24/02/2020.
+//  Copyright © 2020 Tobi Kuyoro. All rights reserved.
+//
+
+import UIKit
+
+class EntryDetailViewController: UIViewController {
+    
+    // MARK: - Outlets
+    
+    @IBOutlet weak var titleTextField: UITextField!
+    @IBOutlet weak var bodyTextView: UITextView!
+    
+    // MARK: - View Lifecycle
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+    // MARK: - Actions
+    
+    @IBAction func save(_ sender: Any) {
+        
+    }
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
