+++ Journal/Journal/EntryTableViewCell.swift
@@ -1,0 +1,30 @@
+//
+//  EntryTableViewCell.swift
+//  Journal
+//
+//  Created by Tobi Kuyoro on 24/02/2020.
+//  Copyright © 2020 Tobi Kuyoro. All rights reserved.
+//
+
+import UIKit
+
+class EntryTableViewCell: UITableViewCell {
+    
+    // MARK: - Outlets
+    
+    @IBOutlet weak var titleLabel: UILabel!
+    @IBOutlet weak var bodyTextLabel: UILabel!
+    @IBOutlet weak var dateLabel: UILabel!
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+
+}
