--- Journal/Journal/Models/CoreDataTask.swift
@@ -1,30 +1,0 @@
-//
-//  CoreDataTask.swift
-//  Journal
-//
-//  Created by Tobi Kuyoro on 27/01/2020.
-//  Copyright © 2020 Tobi Kuyoro. All rights reserved.
-//
-
-import Foundation
-import CoreData
-
-class CoreDataTask {
-    
-    static let shared = CoreDataTask()
-    private init() {}
-    
-    lazy var container: NSPersistentContainer = {
-        let container = NSPersistentContainer(name: "Entries")
-        container.loadPersistentStores { (_, error) in
-            if let error = error {
-                fatalError("Failed to load persistent stores: \(error)")
-            }
-        }
-        return container
-    }()
-    
-    var mainContext: NSManagedObjectContext {
-        return container.viewContext
-    }
-}
