--- Journal/Journal/Models/Entry+Convenience.swift
@@ -1,20 +1,0 @@
-//
-//  Entry+Convenience.swift
-//  Journal
-//
-//  Created by Tobi Kuyoro on 27/01/2020.
-//  Copyright © 2020 Tobi Kuyoro. All rights reserved.
-//
-
-import Foundation
-import CoreData
-
-extension Entry {
-    convenience init(title: String, bodyText: String, timeStamp: Date, identifier: String, context: NSManagedObjectContext) {
-        self.init(context: context)
-        self.title = title
-        self.bodyText = bodyText
-        self.timeStamp = timeStamp
-        self.identifier = identifier
-    }
-}
