+++ Journal/Journal/SceneDelegate.swift
@@ -2,7 +2,7 @@
 //  SceneDelegate.swift
 //  Journal
 //
-//  Created by Tobi Kuyoro on 27/01/2020.
+//  Created by Tobi Kuyoro on 24/02/2020.
 //  Copyright © 2020 Tobi Kuyoro. All rights reserved.
 //
